Improved Version Compatibility in PortalCreation
This update introduces a strategic enhancement within the PortalCreation class to ensure seamless operation across a wider range of Minecraft server versions. Specifically, we've implemented a fallback mechanism within the `catch` block of the PortalCreation class. When an exception is caught — indicating the absence or incompatibility of the BlockData API in the server version being used — the plugin now gracefully falls back to using the PortalCreationLegacy class. This change guarantees that portals are correctly oriented even in environments running older versions of Minecraft, thereby enhancing the plugin's versatility and reliability across diverse server setups. This adjustment is a testament to our dedication to making Iportal as accessible and usable as possible, irrespective of the server version in use.