Portal Creation Process Optimization to Reduce Lag
In this update, I have taken a significant step to improve the performance of the Iportal plugin during portal creation. We addressed an inefficient practice where the execution of both modern and legacy portal creation logics led to unnecessary lag due to duplicate execution. The new implementation aims to eliminate this duplication and optimize execution based on server version.

Changes:

The previous approach of directly invoking PortalCreation.run() regardless of the availability of the BlockData API, followed by a fallback to PortalCreationLegacy.run() in case of failure, resulted in double processing and lag. This method has been removed.

We've implemented a more efficient API availability check before executing the portal creation logic. With a try-catch block, we now first check for the existence of the BlockData class.

This change ensures that based on the availability of the BlockData API, either the modern or the legacy creation logic is executed only once, thus avoiding the previously experienced lag.
With this update, we ensure efficient and lag-free portal creation, improve server performance, and maintain broad compatibility across various Minecraft versions.

Developed by JuL1En to significantly enhance performance while also increasing the plugin's compatibility.