Integrated LanguageManager and Enhancements to Plugin Initialization
This commit introduces significant enhancements and new features to the IPortal plugin to improve user experience and developer friendliness. The `IP` main class has been updated to integrate multilingual support through the `LanguageManager` and to enhance logging upon plugin startup.

Key Changes:

- **Integration of LanguageManager**: The `LanguageManager` class has been added and integrated into the `onEnable` method of the `IP` class. This allows the plugin to support localized messages, facilitating adaptation to various languages.

- **Enhanced Logging at Startup**: The `onEnable` method has been expanded to display an ASCII art logo and information about the plugin version and authors in the server log. This not only provides aesthetic enhancements but also clear information about the plugin at startup.

- **Dynamic Plugin Initialization Notifications**: In addition to existing messages announcing the plugin's activation and deactivation, dynamic information about the plugin's version and contributors is now displayed in the console log. This promotes transparency and makes it easier to identify the plugin version and recognize the contributors.

- **Listing of Authors**: An explicit listing of authors, including the original author, resource manager, and the developer who has been updating the plugin since version 2.0, has been added to acknowledge each contributor's effort.

These changes not only enhance the functionality and usability of the plugin but also contribute to the maintenance of the plugin community by highlighting the significant contributions of developers and contributors.

Implemented by _JuL1En_, this commit marks a significant milestone in the continued development of the IPortal plugin by laying a strong foundation for future expansions and improvements.